% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Function}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Character string. The name to greet. Default is "world".}
}
\value{
A character string with a greeting message.
}
\description{
A simple example function that returns a greeting message.
This demonstrates basic package functionality.
}
\examples{
hello()
hello("imputation-qc")

}
