% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationqc-package.R
\docType{package}
\name{imputationqc-package}
\alias{imputationqc}
\alias{imputationqc-package}
\title{imputationqc: Quality Control Tools for Imputation}
\description{
Provides tools for quality control of genetic imputation results.
    This package helps researchers validate and assess the quality of imputed
    genotype data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pixel-Dream/imputation-qc_dev}
}

}
\author{
\strong{Maintainer}: Maintainer Name \email{maintainer@example.com}

}
\keyword{internal}
